move pkg/p -> pkg/schemes
