+++ pkg/owl/channel.go
@@ -15,9 +15,9 @@ import (
 	"github.com/owlmessenger/owl/pkg/cflog"
 	"github.com/owlmessenger/owl/pkg/dbutil"
 	"github.com/owlmessenger/owl/pkg/owldag"
-	"github.com/owlmessenger/owl/pkg/p/contactset"
-	"github.com/owlmessenger/owl/pkg/p/directmsg"
-	"github.com/owlmessenger/owl/pkg/p/directory"
+	"github.com/owlmessenger/owl/pkg/schemes/contactset"
+	"github.com/owlmessenger/owl/pkg/schemes/directmsg"
+	"github.com/owlmessenger/owl/pkg/schemes/directory"
 	"github.com/owlmessenger/owl/pkg/slices2"
 )
